# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the rest-api documentation",
        "title": "rest-api DOC",
        "contact": {},
        "version": "beta"
    },
    "host": "localhost:7010",
    "paths": {
        "/api/book": {
            "post": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "Create new book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create new book",
                "parameters": [
                    {
                        "description": "book data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateBookInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/book/:id": {
            "get": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "Find a book",
                "produces": [
                    "application/json"
                ],
                "summary": "Find a book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "book identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "Delete a book",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "book identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "Update a book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "book identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "book data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateBookInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/books": {
            "get": {
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "description": "Get all books",
                "produces": [
                    "application/json"
                ],
                "summary": "Get all books",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CreateBookInput": {
            "type": "object",
            "required": [
                "author",
                "author_id",
                "title"
            ],
            "properties": {
                "author": {
                    "type": "string"
                },
                "author_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Authorization": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}